// apps/amplify-be/src/graphql/API.ts
export type User = {
  __typename: "User",
  username: string,
  email: string,
  profile?: Profile | null,
  notifications?: ModelNotificationConnection | null,
  createdAt: string,
  updatedAt: string,
  userProfileUsername?: string | null,
};

export type Profile = {
  __typename: "Profile",
  username: string,
  bio?: string | null,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  username: string,
  title: string,
  message: string,
  read: boolean,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
  userNotificationsUsername?: string | null,
};